Fixed pet skill TP assignment.
Before this, the pet's mobskill was assigning it's TP to the pet's TP value after this value had already been assigned to 0.

In addition, many of the avatar bloodpacts were referencing pet:getTP() when this had already been reduced to 0 after preparing the skill. I changed this to skill:getTP() which should now reflect the actual amount of TP spent.